+++ packages/zeebe-element-templates-json-schema/test/fixtures/execution-listener/valid.js
@@ -1,0 +1,42 @@
+export const template = {
+  'name': 'ExecutionListener',
+  'id': 'com.camunda.example.ExecutionListener',
+  'appliesTo': [
+    'bpmn:Task'
+  ],
+  'entriesVisible': {
+    'executionListeners': false
+  },
+  'properties': [
+    {
+      'type': 'Hidden',
+      'value': 'my-listener-type',
+      'binding': {
+        'type': 'zeebe:executionListener',
+        'eventType': 'start'
+      }
+    },
+    {
+      'type': 'Hidden',
+      'value': 'my-listener-type-end',
+      'binding': {
+        'type': 'zeebe:executionListener',
+        'eventType': 'end',
+        'retries': '3'
+      }
+    },
+    {
+      'type': 'Hidden',
+      'generatedValue': {
+        'type': 'uuid'
+      },
+      'binding': {
+        'type': 'zeebe:executionListener',
+        'eventType': 'end',
+        'retries': '3'
+      }
+    }
+  ]
+};
+
+export const errors = null;