+++ packages/zeebe-element-templates-json-schema/test/fixtures/task-listener/invalid-missing-type.js
@@ -1,0 +1,58 @@
+export const template = {
+  'name': 'test',
+  'id': 'test',
+  'appliesTo': [
+    'bpmn:Task'
+  ],
+  'elementType': {
+    'value': 'bpmn:UserTask'
+  },
+  'properties': [
+    {
+      'value': 'my-listener-type',
+      'binding': {
+        'type': 'zeebe:taskListener',
+        'eventType': 'assigning'
+      }
+    }
+  ]
+};
+
+export const errors = [
+  {
+    keyword: 'required',
+    dataPath: '/properties/0',
+    schemaPath: '#/allOf/1/items/allOf/29/then/required',
+    params: {
+      missingProperty: 'type'
+    },
+    message: 'should have required property \'type\''
+  },
+  {
+    keyword: 'if',
+    dataPath: '/properties/0',
+    schemaPath: '#/allOf/1/items/allOf/29/if',
+    params: {
+      failingKeyword: 'then'
+    },
+    message: 'should match "then" schema'
+  },
+  {
+    keyword: 'type',
+    dataPath: '',
+    schemaPath: '#/oneOf/1/type',
+    params: {
+      type: 'array'
+    },
+    message: 'should be array'
+  },
+  {
+    keyword: 'oneOf',
+    dataPath: '',
+    schemaPath: '#/oneOf',
+    params: {
+      passingSchemas: null
+    },
+    message: 'should match exactly one schema in oneOf'
+  }
+];