+++ packages/zeebe-element-templates-json-schema/CHANGELOG.md
@@ -6,6 +6,11 @@ All notable changes to [@camunda/zeebe-element-templates-json-schema](https://gi
 
 ___Note:__ Yet to be released changes appear here._
 
+## 0.38.0
+
+* `FEAT`: add schema support of templating execution and task listeners ([#224](https://github.com/camunda/element-templates-json-schema/pull/224))
+* `FEAT`: allow controlling visibility of an execution and task listeners section via `entriesVisible` ([#224](https://github.com/camunda/element-templates-json-schema/pull/224))
+
 ## 0.37.0
 
 * `FEAT`: allow `entriesVisible` to accept an object with property group ids to show them in properties panel with applied template ([#218](https://github.com/camunda/element-templates-json-schema/pull/218))